refactor: move export_jsonlines to exporters.py per project requirements
