+++ pytrackio/_report.py
@@ -33,36 +33,6 @@ def export_json(path=None,*,indent=2):
         with open(path,"w") as f: f.write(d)
     return d
 
-def export_jsonlines(path=None):
-    """
-    Exports metrics (summaries + counters) in JSON Lines format.
-    Each line is a valid standalone JSON object.
-    """
-    output = []
-    summaries = _REGISTRY.all_summaries()
-    counters = _REGISTRY.all_counters()
-    for s in summaries:
-        obj = {
-            "name": s.name,
-            "calls": s.calls,
-            "avg": round(s.avg_ms, 2),
-            "min": round(s.min_ms, 2),
-            "max": round(s.max_ms, 2),
-            "p95": round(s.p95_ms, 2),
-            "p99": round(s.p99_ms, 2),
-            "errors": s.errors
-        }
-        output.append(json.dumps(obj)) 
-
-    for name, value in counters.items():
-        obj = {"name": name, "type": "counter", "value": value}
-        output.append(json.dumps(obj))
-    result = "\n".join(output)
-    if path:
-        with open(path, "w") as f:
-            f.write(result + "\n")
-    return result
-
 def export_csv(path=None):
     buf=io.StringIO(); fields=["name","calls","errors","error_rate","avg_ms","min_ms","max_ms","p95_ms","p99_ms"]
     w=csv.DictWriter(buf,fieldnames=fields); w.writeheader()
